# keep cache and constraint store after recursion, revert state on exceptions