… booster/test/rpc-integration/test-log-simplify-json/simplify-log.txt.golden