Previously we were discarding any simplifications cached during recursive simplification. 
This is overly cautious - while the change flag and the term stack (for iteration count and loop detection) should be restored, the cache remains valid (we may only _gain_ new path conditions, and only in rare cases where equations have an `ensures` clause), and continues to be useful when the same or a similar recursive evaluation is needed.

This change sped up a particular request extracted from an `mx-backend` proof from 5 minutes to 5 seconds.